+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 Change Log
 ==========
 
+### Features
+* Cria relatório de CPK por marca, modelo e dimensão de pneu (PL-2699)
+
 <a name="v1.0.31"></a>
 ## Version [v1.0.31](https://github.com/luizfp/PrologWebService/compare/v1.0.30...v1.0.31) (release-date) [unreleased]
 
 ### Bug Fixes
 * Corrige busca de modelos de quizzes para realização
-
 
 <a name="v1.0.30"></a>
 ## Version [v1.0.30](https://github.com/luizfp/PrologWebService/compare/v1.0.29...v1.0.30) (release-date) [unreleased]